+++ core/src/test/java/io/honeycomb/opentelemetry/android/HoneycombOptionsUnitTest.kt
@@ -879,4 +879,26 @@ class HoneycombOptionsUnitTest {
             HoneycombOptions.Builder(source).build()
         }
     }
+
+    @Test
+    fun options_throwsOnMissingApiKey() {
+        val source =
+            HoneycombOptionsMapSource(
+                mapOf<String, Any?>(),
+            )
+        assertThrows(HoneycombException::class.java) {
+            HoneycombOptions.Builder(source).build()
+        }
+    }
+
+    @Test
+    fun options_noThrowsOnMissingApiKeyWithCustomEndpoint() {
+        val source =
+            HoneycombOptionsMapSource(
+                mapOf<String, Any?>(
+                    "HONEYCOMB_API_ENDPOINT" to "http://example.com:1234",
+                ),
+            )
+        HoneycombOptions.Builder(source).build()
+    }
 }